flytt søknad typer sammen
