+++ apps/foreldrepengesoknad/src/api/api.ts
@@ -3,14 +3,13 @@ import { AxiosResponse } from 'axios';
 import Fordeling from 'types/Fordeling';
 import { Kvittering } from 'types/Kvittering';
 import { SakerOppslag } from 'types/SakerOppslag';
-import { Søknad } from 'types/Søknad';
+import { EndringssøknadForInnsending, Søknad, SøknadForInnsending } from 'types/Søknad';
 import { useGetRequest } from 'utils/hooks/useRequest';
 
 import { getAxiosInstance } from '@navikt/fp-api';
 import { BarnFraNesteSak, EksisterendeSak, Periode } from '@navikt/fp-common';
 import { Attachment, Dekningsgrad, LocaleNo, Søkerinfo } from '@navikt/fp-types';
 
-import { EndringssøknadForInnsending, SøknadForInnsending } from './apiUtils';
 import { storageParser } from './storageParser';
 
 export interface TilgjengeligeStønadskontoerParams {
